Box whitespace must be minimum not maximum of child whitespaces
Because it's the distance from the outermost pixels to the edge
of the box. e.g. if one child in a column has a left axis and the
other doesn't and the plot areas are aligned, the one with the left
axis has less whitespace and defines the box whitespace.